Revert "Back commonPipelineEnvironment step by shared class"
This reverts commit 149cd96dbfab3374ae20182d570fd1f0233014f2.
